+++ src/views/about_view.py
@@ -28,7 +28,7 @@ class AboutView(BaseView):
 
         info = info_label(
             container,
-            "CoolBox - A Modern Desktop App\nVersion 1.0.59",
+            "CoolBox - A Modern Desktop App\nVersion 1.0.60",
             font=self.font,
         )
         info.pack(anchor="w")